Temporary workaround for very small invalid k(E) values
Until we figure out the actual problem with #77, this fix should
allow RMG runs to continue if a very small invalid k(E) value is
generated. The error is still reported though, both in the log
and in the reaction's kinetics comments.

The cutoff is set to 1e-35 for now since the largest invalid k(E)
we've reported is on the order of 1e-40, but can be easily
modified as necessary.